// src/app/store/model/events/events.enum.ts
export enum CandidateEventType {
  SESSION_STARTED = 'SESSION_STARTED',
  SESSION_RESUMED = 'SESSION_RESUMED',
  SESSION_PAUSED = 'SESSION_PAUSED',
  SESSION_TERMINATED = 'SESSION_TERMINATED',
  SESSION_TIMEOUT = 'SESSION_TIMEOUT',
  SESSION_LOGOUT = 'SESSION_LOGOUT',

  LOGIN_ATTEMPTED = 'LOGIN_ATTEMPTED',
  LOGIN_SUCCEEDED = 'LOGIN_SUCCEEDED',
  LOGIN_FAILED = 'LOGIN_FAILED',
  SECURE_ENVIRONMENT_FAILED = 'SECURE_ENVIRONMENT_FAILED',

  ONBOARDING_STARTED = 'ONBOARDING_STARTED',
  ONBOARDING_STEP_CHANGED = 'ONBOARDING_STEP_CHANGED',
  SYSTEM_CHECK_STARTED = 'SYSTEM_CHECK_STARTED',
  SYSTEM_CHECK_COMPLETED = 'SYSTEM_CHECK_COMPLETED',
  SYSTEM_CHECK_FAILED = 'SYSTEM_CHECK_FAILED',
  MIC_TEST_STARTED = 'MIC_TEST_STARTED',
  MIC_TEST_COMPLETED = 'MIC_TEST_COMPLETED',
  CAMERA_TEST_STARTED = 'CAMERA_TEST_STARTED',
  CAMERA_TEST_COMPLETED = 'CAMERA_TEST_COMPLETED',
  GUIDELINES_ACCEPTED = 'GUIDELINES_ACCEPTED',

  INSTRUCTIONS_VIEWED = 'INSTRUCTIONS_VIEWED',
  READING_TIME_WARNING = 'READING_TIME_WARNING',
  READING_TIME_EXPIRED = 'READING_TIME_EXPIRED',
  EXAM_START_REQUESTED = 'EXAM_START_REQUESTED',
  EXAM_SELECTED = 'EXAM_SELECTED',
  EXAM_STARTED = 'EXAM_STARTED',
  EXAM_SUBMIT_REQUESTED = 'EXAM_SUBMIT_REQUESTED',
  EXAM_SUBMIT_CANCELLED = 'EXAM_SUBMIT_CANCELLED',
  EXAM_SUBMITTED = 'EXAM_SUBMITTED',
  EXAM_AUTO_SUBMITTED = 'EXAM_AUTO_SUBMITTED',

  SECTION_ENTERED = 'SECTION_ENTERED',
  SECTION_EXITED = 'SECTION_EXITED',

  QUESTION_ENTERED = 'QUESTION_ENTERED',
  QUESTION_EXITED = 'QUESTION_EXITED',
  ANSWER_SELECTED = 'ANSWER_SELECTED',
  ANSWER_CHANGED = 'ANSWER_CHANGED',
  ANSWER_CLEARED = 'ANSWER_CLEARED',
  QUESTION_FLAGGED = 'QUESTION_FLAGGED',
  QUESTION_UNFLAGGED = 'QUESTION_UNFLAGGED',

  EXAM_OVERVIEW_OPENED = 'EXAM_OVERVIEW_OPENED',
  LAYOUT_CHANGED = 'LAYOUT_CHANGED',
  INSTRUCTIONS_PANEL_TOGGLED = 'INSTRUCTIONS_PANEL_TOGGLED',
  PROFILE_OPENED = 'PROFILE_OPENED',
  CALCULATOR_OPENED = 'CALCULATOR_OPENED',
  CALCULATOR_CLOSED = 'CALCULATOR_CLOSED',

  NETWORK_ONLINE = 'NETWORK_ONLINE',
  NETWORK_OFFLINE = 'NETWORK_OFFLINE',
  SYNC_STARTED = 'SYNC_STARTED',
  SYNC_COMPLETED = 'SYNC_COMPLETED',
  SYNC_FAILED = 'SYNC_FAILED',

  WINDOW_BLURRED = 'WINDOW_BLURRED',
  WINDOW_FOCUSED = 'WINDOW_FOCUSED',
  COPY_ATTEMPT = 'COPY_ATTEMPT',
  PASTE_ATTEMPT = 'PASTE_ATTEMPT',
  PRINT_ATTEMPT = 'PRINT_ATTEMPT',
  SCREENSHOT_ATTEMPT = 'SCREENSHOT_ATTEMPT',
  FULLSCREEN_ENTERED = 'FULLSCREEN_ENTERED',
  FULLSCREEN_EXITED = 'FULLSCREEN_EXITED',

  BATTERY_LOW = 'BATTERY_LOW',
  DEVICE_SLEEP = 'DEVICE_SLEEP',
  DEVICE_RESUMED = 'DEVICE_RESUMED',

  PROCTOR_CHAT_OPENED = 'PROCTOR_CHAT_OPENED',
  PROCTOR_CHAT_CLOSED = 'PROCTOR_CHAT_CLOSED',
  PROCTOR_MESSAGE_SENT = 'PROCTOR_MESSAGE_SENT',
  PROCTOR_MESSAGE_RECEIVED = 'PROCTOR_MESSAGE_RECEIVED',

  FACE_DETECTED = 'FACE_DETECTED',
  FACE_LOST = 'FACE_LOST',
  MULTIPLE_FACES = 'MULTIPLE_FACES',
  PHONE_DETECTED = 'PHONE_DETECTED',
  BOOK_DETECTED = 'BOOK_DETECTED',
  LOOKING_AWAY = 'LOOKING_AWAY',
  MIC_BLOCKED = 'MIC_BLOCKED',
  CAMERA_BLOCKED = 'CAMERA_BLOCKED',

  APPLICATION_EXIT_REQUESTED = 'APPLICATION_EXIT_REQUESTED',
  APPLICATION_EXITED = 'APPLICATION_EXITED',
  APPLICATION_CRASHED = 'APPLICATION_CRASHED',
}

export enum NavigationMethod {
  Next = 'Next',
  Prev = 'Prev',
  Palette = 'Palette',
  Keyboard = 'Keyboard',
}